requirements.txt
streamlit==1.28.1
torch==2.0.1
torchvision==0.15.2
onnxruntime==1.16.0
tensorflow==2.13.0
adversarial-robustness-toolbox==1.14.0
numpy==1.24.3
pillow==10.0.1
opencv-python==4.8.1.78
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
reportlab==4.0.4
scikit-learn==1.3.0
scipy==1.11.3
requests==2.31.0
cryptography==41.0.3





